feat: Inject TPU envvars into StatefulSet and pods
This is a bare bones implementation of doing the env var injection
required by TPUs to be able to function correctly across TPU pod slices.

This doesn't include any kind of error handling or any good logging
either.

However, if you create a statefulset and annotate the statefulset and
the pods with the right annotation, then you get both the
`TPU_WORKER_ID` based on the StatefulSet ordinal and you get the
`TPU_WORKER_HOSTNAMES` which the `tpu_driver` can use for setting up
multi host pod slices.